+++ crates/sui-core/src/authority_aggregator.rs
@@ -389,7 +389,7 @@ where
         FReduce: Fn(
             S,
             AuthorityName,
-            usize,
+            u64,
             Result<V, SuiError>,
         ) -> AsyncResult<'a, ReduceOutput<S>, SuiError>,
     {
@@ -466,8 +466,8 @@ where
     > {
         #[derive(Default)]
         struct GetObjectByIDRequestState {
-            good_weight: usize,
-            bad_weight: usize,
+            good_weight: u64,
+            bad_weight: u64,
             responses: Vec<(AuthorityName, SuiResult<ObjectInfoResponse>)>,
         }
         let initial_state = GetObjectByIDRequestState::default();
@@ -613,8 +613,8 @@ where
     ) -> Result<(BTreeMap<ObjectRef, Vec<AuthorityName>>, Vec<AuthorityName>), SuiError> {
         #[derive(Default)]
         struct OwnedObjectQueryState {
-            good_weight: usize,
-            bad_weight: usize,
+            good_weight: u64,
+            bad_weight: u64,
             object_map: BTreeMap<ObjectRef, Vec<AuthorityName>>,
             responded_authorities: Vec<AuthorityName>,
             errors: Vec<(AuthorityName, SuiError)>,
@@ -877,8 +877,8 @@ where
             // The list of errors gathered at any point
             errors: Vec<SuiError>,
             // Tally of stake for good vs bad responses.
-            good_stake: usize,
-            bad_stake: usize,
+            good_stake: u64,
+            bad_stake: u64,
         }
 
         let state = ProcessTransactionState {
@@ -1034,8 +1034,8 @@ where
             // Different authorities could return different effects.  We want at least one effect to come
             // from 2f+1 authorities, which meets quorum and can be considered the approved effect.
             // The map here allows us to count the stake for each unique effect.
-            effects_map: HashMap<[u8; 32], (usize, TransactionEffects)>,
-            bad_stake: usize,
+            effects_map: HashMap<[u8; 32], (u64, TransactionEffects)>,
+            bad_stake: u64,
         }
 
         let state = ProcessCertificateState {
@@ -1125,7 +1125,7 @@ where
                             let entry = state
                                 .effects_map
                                 .entry(inner_effects.digest())
-                                .or_insert((0usize, inner_effects.effects));
+                                .or_insert((0, inner_effects.effects));
                             entry.0 += weight;
 
                             if entry.0 >= threshold {
@@ -1221,7 +1221,7 @@ where
         while let Some(((obj_ref, tx_digest), (obj_option, layout_option, authorities))) =
             object_ref_stack.pop()
         {
-            let stake: usize = authorities
+            let stake: u64 = authorities
                 .iter()
                 .map(|(name, _)| self.committee.weight(name))
                 .sum();
