+++ crates/sui-types/src/committee.rs
@@ -16,15 +16,15 @@ pub type EpochId = u64;
 #[derive(Eq, PartialEq, Clone, Debug, Serialize, Deserialize)]
 pub struct Committee {
     pub epoch: EpochId,
-    pub voting_rights: BTreeMap<AuthorityName, usize>,
-    pub total_votes: usize,
+    pub voting_rights: BTreeMap<AuthorityName, u64>,
+    pub total_votes: u64,
     // Note: this is a derived structure, no need to store.
     #[serde(skip)]
     pub expanded_keys: HashMap<AuthorityName, PublicKey>,
 }
 
 impl Committee {
-    pub fn new(epoch: EpochId, voting_rights: BTreeMap<AuthorityName, usize>) -> Self {
+    pub fn new(epoch: EpochId, voting_rights: BTreeMap<AuthorityName, u64>) -> Self {
         let total_votes = voting_rights.iter().map(|(_, votes)| votes).sum();
         let expanded_keys: HashMap<_, _> = voting_rights
             .iter()
@@ -57,17 +57,17 @@ impl Committee {
         unreachable!();
     }
 
-    pub fn weight(&self, author: &AuthorityName) -> usize {
+    pub fn weight(&self, author: &AuthorityName) -> u64 {
         *self.voting_rights.get(author).unwrap_or(&0)
     }
 
-    pub fn quorum_threshold(&self) -> usize {
+    pub fn quorum_threshold(&self) -> u64 {
         // If N = 3f + 1 + k (0 <= k < 3)
         // then (2 N + 3) / 3 = 2f + 1 + (2k + 2)/3 = 2f + 1 + k = N - f
         2 * self.total_votes / 3 + 1
     }
 
-    pub fn validity_threshold(&self) -> usize {
+    pub fn validity_threshold(&self) -> u64 {
         // If N = 3f + 1 + k (0 <= k < 3)
         // then (N + 2) / 3 = f + 1 + k/3 = f + 1
         (self.total_votes + 2) / 3
@@ -89,7 +89,7 @@ impl Committee {
     pub fn robust_value<A, V>(
         &self,
         items: impl Iterator<Item = (A, V)>,
-        threshold: usize,
+        threshold: u64,
     ) -> (AuthorityName, V)
     where
         A: Borrow<AuthorityName> + Ord,
